// src/pages/Game/SpinGame/types.d.ts
export type ImageResourcesType = {
  bg: string
  btn: string
  btn_answer: string
  btn_back: string
  btn_null: string
  btn_sound_off: string
  btn_sound_on: string
  btn_summary: string
  game_name: string
  ico_board_correct: string
  ico_board_summary: string
  ico_board_text: string
  ico_heart_empty: string
  ico_heart_full: string
  ico_result_correct: string
  ico_result_wrong: string
  ico_robo_happy: string
  ico_robo_sad: string
  ico_sound: string
  ico_spin: string
  ico_spin_board: string
  ico_spin_board_small: string
  logo: string
}
export type AudioResourcesType = {
  sound_correct: string
  sound_spin: string
  sound_wrong: string
  theme: string
  theme_summary: string
}